Add seeder files, controllers, and update api routes
Added seeder files for the Users, Clubs, Courts, RoleUser tables. Implemented Matchmaking Controller and moved Role, User, Club, and Court Controllers to a `Settings` namespace. Refined the API routes to adjust paths following namespace changes and added matchmaking endpoint. Furthermore, the Users table was updated with an 'age' column and a fix was applied in the RoleResource file to return role ID.