Add SoftDeletes trait to models and migrations
This commit adds the SoftDeletes trait to various models and migrations. It also introduces a new PlayedMatch model and creates the corresponding migration for 'matches' table. For the migrations, 'softDeletes' method is added to maintain records of deleted entries. In models, the 'SoftDeletes' trait is included to apply the necessary soft delete methods.