Add match request and match handling functionality
This commit introduces a new RequestForMatch model and request validation rules, as well as significant changes to the MatchController. It also renames the PlayedMatch model to TennisMatch for clarity. Functionality for request, acceptance, and declining of matches is added, including new routes and controller functions. The MatchRequest model also supports relationship bindings to the TennisMatch model and the User model. "Played_at" field is added to the matches table schema.